+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
+## [0.4.0] - 2026-04-09
+
+### Added
+
+- Rich console output with grouped panels for scan results
+- `scan` shows dirty/ahead/behind/clean status per repo
+- `--no-color` flag to disable colors
+- `--version` flag to print the current version
+- Rich dependency
+
 ## [0.3.0] - 2026-04-05
 
 ### Added
@@ -45,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Documentation files (README, CHANGELOG, CONTRIBUTING)
 
+[0.4.0]: https://gitlab.com/neo-bend-reality/reposweep/-/compare/v0.3.0...HEAD/
 [0.3.0]: https://gitlab.com/neo-bend-reality/reposweep/-/compare/v0.2.0...v0.3.0
 [0.2.0]: https://gitlab.com/neo-bend-reality/reposweep/-/compare/v0.1.0...v0.2.0
 [0.1.0]: https://gitlab.com/neo-bend-reality/reposweep/-/compare/v0.0.2...v0.1.0